cmd/lncli: move invoice commands to cmd_invoice.go
